+++ pkg/daemons/executor/embed.go
@@ -95,7 +95,7 @@ func (e *Embedded) Kubelet(ctx context.Context, args []string) error {
 
 func (e *Embedded) KubeProxy(ctx context.Context, args []string) error {
 	command := proxy.NewProxyCommand()
-	command.SetArgs(daemonconfig.GetArgs(platformKubeProxyArgs(e.nodeConfig), args))
+	command.SetArgs(util.GetArgs(platformKubeProxyArgs(e.nodeConfig), args))
 
 	go func() {
 		<-e.APIServerReadyChan()
